Upgrade dependencies

+++ beeref/assets/__init__.py
@@ -51,4 +51,5 @@ class BeeAssets:
         img = QtGui.QImage(os.path.join(self.PATH, filename))
         pixmap = QtGui.QPixmap.fromImage(img)
         pixmap.setDevicePixelRatio(scaling)
-        return QtGui.QCursor(pixmap, hotspot[0]/scaling, hotspot[1]/scaling)
+        return QtGui.QCursor(
+            pixmap, int(hotspot[0]/scaling), int(hotspot[1]/scaling))
